Add files via upload
truly, we live in a hell of our own creation (i just wanted to have a consistent font for multiple scripts, this is such an ass backwards way to implement this). I've included all the OFL files for it bc that was easier than checking which are identical, but given its just two fonts from Google Fonts I'm in no hot water.